Split test files based on the implementation files
